requirements.txt
asgiref==3.2.7
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
Werkzeug==1.0.1
whitenoise==5.1.0
